CMake: use PACKAGE_VERSION for the pkg-config file
The old variable `OPUS_LIBRARY_VERSION`, does not line up with what's used in autoconf, which uses the PACKAGE_VERSION instead, which in turn lines up with what version the projects `opusfile` and `libsndfile` check for, for example.
While at it, I also cleaned up the accidental double up in code.